Set nO dimension on the transformer model
We use this dimension to create an empty array with the correct width
when a doc is empty. So, we should actually set this dimension.

While making these changes, I also noticed that the CamemBERT factory
did not accept the `mixed_precision` and `grad_scaler_config` arguments,
so I fixed that as well.