Add support for replacing listeners (#7)
* Add support for replacing listeners.

Multiple changes were required to facilitate this:
* All transformer model entrypoints now have a `wrapped_listener` optional parameter.
This parameter is only meant to be used by the machinery that performs the listener
replacement.
* Listeners are no longer subclasses of the `TransformerListener` class. Previously,
the `TransformerListener` class subclassed `Model` and stored some state as instance
attributes. To perform the replacement, the original listener instance in the downstream
component needs to be (deep)copied. However, the implementation of `Model.copy` doesn't
support classes that subclass `Model` - it merely performs deepcopies of the different
`Model` instance attributes and initializes a new `Model` instance with them. What this
results in is the loss of any state that was directly stored on the listener instance such
as `upstream_name`, `name`, etc.

To workaround this limitation, all listener state is now directly stored in `Model.attrs`.
This ensures that no persistent state is lost between copies.
* A new `WrappedTransformerAndListener` class has been introduced to be used as the replacement
model for the downstream component's original listener. This wraps the upstream transformer pipe's
model and the original listener. During training and prediction, it calls the wrapped transformer
and direcly passes the outputs to the wrapped listener. Gradients are additionally allocated during
training, and during prediction, the wrapped listener is instructed to ignore any transformer annotations
present on the `Doc`s and use the ones directly stored in the listener.

* Increment `curated-tokenizers` lowerbound to `0.0.8`

* Increment `spacy` lowerbound to `3.7.0.dev0`

* Restrict `spacy` requirement upperbound

* Fix listener test

* Fix impossible condition in `verify_inputs`

* Raise exception when attempting to register a non-listener model with the pipe

* Don't allocate gradients during inference

* Allow 3rd party listeners to use all transformer layer outputs

* Rename `TransformerListener` to `ListenerStateUtils`

* Switch to free functions for listener `forward` functions
Inline listener construction code and remove classes