Load models in-place (#32)
* Load models in-place

Before this change, we would first construct a model from using the
arguments passed to the registry function. Then we would construct it
again using `from_hf_hub`. This was not only a performance issue, but
also a correctness issues -- the model constructed through `from_hf_hub`
could have different hyperparameters than those specified in the
arguments to the registry function.

This change fixed this by using the new in-place loading support in
Curated Transformers 2.0.

The addition to in-place loading also added the `dtype` argument to the
model configuration. We also expose this argument now in v2 versions of
the registry functions. The configuration filling is also updated to
fill the data type from the `torch_dtype` option in the HF model
configuration.

* Loosen atol a bit on XLM-R pipeline test

Otherwise fails on GPU, but the difference is too
small to make a difference.