Make 'whitespace alignment' a no-op (#39)
A vector is being inserted for each whitespace token, which requires somewhat expensive copying of the data. Instead, we can rely on the fact that the pooling operations that calculate vectors for each token in the models support 0-length sequences in the Ragged type.

In more detail: The output types are ragged arrays that effectively represent nested variable-length sequences -- like a list of arrays. Each item in the 'list' is an array of wordpieces aligned to a spaCy token.

The whitespace tokens in spaCy are stripped before the data is passed to the wordpiece tokenization, so we have no wordpieces aligned to the whitespace. Before this patch we handled this by recomputing the outputs with new rows, representing these whitespace tokens. What we do now is simply let them be a 0-length sequence.

This change may affect users who directly access the doc._.trf_data structure, as a doc._.trf_data[i] index might now retrieve a zero-length array, where before it retrieved a zeroed array.

I see this potential incompatibility as an acceptable risk to getting the ~20% performance benefit from this change. I expect relatively few users to be accessing the arrays directly, and a conditional check of the length can be easily inserted to account for the whitespace regions. It isn't a bad thing to handle the whitespace regions explicitly, and it's also an advantage that we're not changing the output tensors from the underlying transformer, making it easier to interact with the underlying model.

See this thread for more detailed discussion: explosion/spaCy#13783